Added UEM compatible mode switch
Added folder removing cmd to watchdog in lieu of UEM switch
Added timestamp to console messages
Removed unnecessary output

+++ silent-tester-runner-windows.ps1
@@ -8,7 +8,7 @@ param (
   [int] $ScenarioDuration = 86400, # defaults to 24 hours - can be changed
   [Parameter(Mandatory=$false)]
   [Alias ("SCCM","Intune")]
-  [switch] $EM_Comatible_Mode,
+  [switch] $UEM_Comatible_Mode,
   [Parameter(Mandatory=$false)]
   [Alias ("Force")]
   [switch] $AllowMultipleRuns, # Use if you want to be able to run more the once with the same $TestID on the machine. 2 tabs might jump to user.
@@ -23,6 +23,7 @@ $p5UserDataDir = "$env:TEMP\p5-user-" + $TestID
 $preferencesFilePath = $p5UserDataDir + "\Default\Preferences"
 $cacheFolderPath = $p5UserDataDir + "\Default\Cache"
 $defaultPaths = @("C:\Program Files (x86)\Microsoft\Edge\Application\msedge.exe","C:\Program Files (x86)\Google\Chrome\Application\chrome.exe", "C:\Program Files\Google\Chrome\Application\chrome.exe")
+$durationMinimum = 60
 
 ### Parameter validation ###
 $RegexForTenantId = '[a-z0-9]{8}\-[a-z0-9]{4}\-[a-z0-9]{4}\-[a-z0-9]{4}\-[a-z0-9]{12}'
@@ -30,8 +31,8 @@ if ($TenantID -notmatch $RegexForTenantId) {
   Write-Error "Invalid Parameter: Tenant ID. Please provide a valid Tenant ID."
   Exit 1
 }
-if ($ScenarioDuration -lt 300) {
-  Write-Error "Invalid Parameter: Scenario Duration. Please provide a Scenario Duration of greater than 300."
+if ($ScenarioDuration -lt $durationMinimum) {
+  Write-Error "Invalid Parameter: Scenario Duration. Please provide a Scenario Duration of greater than $durationMinimum seconds."
   Exit 1
 }
 if ((Test-Path $logPath) -and (!$AllowMultipleRuns)) {
@@ -77,21 +78,23 @@ if (!$CustomChromePath -or !(Test-Path $CustomChromePath)) {
     Exit 2
   }
 }
-Write-Host "Found Chrome path $($CustomChromePath)"
+Write-Host "Found Chrome path '$CustomChromePath'"
 $Process = Start-Process -RedirectStandardOutput $logPath -RedirectStandardError $errLogPath -passthru $CustomChromePath -ArgumentList "$($pageURL)  --hide-crash-restore-bubble --autoplay-policy=no-user-gesture-required --disable-backgrounding-occluded-windows --disable-background-media-suspend --disable-renderer-backgrounding --disable-gpu --remote-debugging-port=0 --disable-infobars --disable-restore-session-state --user-data-dir=$p5UserDataDir --disable-gesture-requirement-for-media-playback --disable-background-networking --disable-background-timer-throttling --disable-breakpad --disable-client-side-phishing-detection --disable-default-apps --disable-dev-shm-usage --disable-extensions --disable-field-trial-config --disable-features=site-per-process,WebRtcHideLocalIpsWithMdns --disable-hang-monitor --disable-popup-blocking --disable-prompt-on-repost --disable-sync --disable-translate --metrics-recording-only --no-first-run --safebrowsing-disable-auto-update --enable-automation --password-store=basic --use-mock-keychain --mute-audio --process-per-site" -WorkingDirectory $env:TEMP
 if ([System.Security.Principal.WindowsIdentity]::GetCurrent().Name -ne "NT AUTHORITY\SYSTEM") {
   While ($Process.MainWindowHandle -eq 0) { Start-Sleep -m 100 }
   [my.WinApi]::Hide($Process.MainWindowHandle)
 }
-Write-Host "Started Chromium process, with id: $($Process.id)"
+Write-Host "$(Get-Date) Started Chromium process, with id: $($Process.id)"
 $chromePid = $Process.id
 $cmd = "cmd.exe"
 $extraTimeout = $ScenarioDuration + 10
-$argos =  "/c timeout ${extraTimeout} && taskkill.exe /f /t /pid ${chromePid}"
+$argos =  "/c timeout ${extraTimeout} && taskkill.exe /f /t /pid $chromePid && rd /s /q $cacheFolderPath"
 $watchdogProcess = Start-Process -WindowStyle hidden -passthru $cmd -ArgumentList $argos
+if ($UEM_Comatible_Mode) {
+  return
+}
 Start-Sleep -s $ScenarioDuration
 $stopProcessInfo = Stop-Process -InputObject $Process -passthru
-$stopProcessInfo
 if (Test-Path $preferencesFilePath) {
   try {
     $Prefs = ((Get-Content $preferencesFilePath) -replace "`"exit_type`":`"Crashed`"" , "`"exit_type`":`"none`"") -replace "`"exited_cleanly`":false","`"exited_cleanly`":true"
@@ -104,4 +107,4 @@ if (Test-Path $cacheFolderPath) {
     Remove-Item -Recurse $cacheFolderPath
   } catch {}
 }
-Write-Host "Stopped Chromium process"
+Write-Host "$(Get-Date) Stopped Chromium process"
